scripts/affiliate/requirements.txt
python-dotenv==1.0.0
requests==2.31.0
firebase-admin==6.3.0
python-amazon-paapi==5.1.1
lxml==4.9.3
beautifulsoup4==4.12.2
